src/components/app-todos/app-todos.spec.ts
import { AppTodos } from './app-todos';

describe('app-todos', () => {
  it('builds', () => {
    expect(new AppTodos()).toBeTruthy();
  });

  describe('normalization', () => {
    it('returns a blank string if the name is undefined', () => {
      const component = new AppTodos();
      expect(component.normalize(undefined)).toEqual('');
    });

    it('returns a blank string if the name is null', () => {
      const component = new AppTodos();
      expect(component.normalize(null)).toEqual('');
    });

    it('capitalizes the first letter', () => {
      const component = new AppTodos();
      expect(component.normalize('quincy')).toEqual('Quincy');
    });

    it('lower-cases the following letters', () => {
      const component = new AppTodos();
      expect(component.normalize('JOSEPH')).toEqual('Joseph');
    });

    it('handles single letter names', () => {
      const component = new AppTodos();
      expect(component.normalize('q')).toEqual('Q');
    });
  });
});
